Hello everyone! This is the first commit for BrotherHoods plugin. We're excited to share our project with you and hope it brings an enjoyable experience for Minecraft players.
This commit sets up the foundation for the plugin, including the creation of factions (or kingdoms), as well as a working command for listing all the kingdoms in the game.

To list all kingdoms, simply type "/kingdoms" in the chat or an alias. We'll be adding more features and improvements as we go, so stay tuned for updates. Thanks for checking out BrotherHoods!